+++ tests/backends/database/test_mongo.py
@@ -1,12 +1,17 @@
 """Tests for Ralph mongo database backend"""
 
+from datetime import datetime
+
 import pytest
 from bson.objectid import ObjectId
 from pymongo import MongoClient
-from pymongo.errors import BulkWriteError
 
 from ralph.backends.database.mongo import MongoDatabase, MongoQuery
-from ralph.exceptions import BackendParameterException, BadFormatException
+from ralph.exceptions import (
+    BackendException,
+    BackendParameterException,
+    BadFormatException,
+)
 
 from tests.fixtures.backends import (
     MONGO_TEST_COLLECTION,
@@ -36,7 +41,13 @@ def test_backends_database_mongo_get_method(mongo):
     """Test the mongo backend get method."""
 
     # Create records
-    documents = MongoDatabase.to_documents([{"id": "foo"}, {"id": "bar"}])
+    timestamp = {"timestamp": "2022-06-27T15:36:50"}
+    documents = MongoDatabase.to_documents(
+        [
+            {"id": "foo", **timestamp},
+            {"id": "bar", **timestamp},
+        ]
+    )
     database = getattr(mongo, MONGO_TEST_DATABASE)
     collection = getattr(database, MONGO_TEST_COLLECTION)
     collection.insert_many(documents)
@@ -48,8 +59,8 @@ def test_backends_database_mongo_get_method(mongo):
         collection=MONGO_TEST_COLLECTION,
     )
     expected = [
-        {"_id": "2c26b46b68ffc68ff99b453c", "_source": {"id": "foo"}},
-        {"_id": "fcde2b2edba56bf408601fb7", "_source": {"id": "bar"}},
+        {"_id": "62b9ce922c26b46b68ffc68f", "_source": {"id": "foo", **timestamp}},
+        {"_id": "62b9ce92fcde2b2edba56bf4", "_source": {"id": "bar", **timestamp}},
     ]
     assert list(backend.get()) == expected
     assert list(backend.get(chunk_size=1)) == expected
@@ -60,8 +71,13 @@ def test_backends_database_mongo_get_method_with_a_custom_query(mongo):
     """Test the mongo backend get method with a custom query."""
 
     # Create records
+    timestamp = {"timestamp": datetime.now().isoformat()}
     documents = MongoDatabase.to_documents(
-        [{"id": "foo", "bool": 1}, {"id": "bar", "bool": 0}, {"id": "lol", "bool": 1}]
+        [
+            {"id": "foo", "bool": 1, **timestamp},
+            {"id": "bar", "bool": 0, **timestamp},
+            {"id": "lol", "bool": 1, **timestamp},
+        ]
     )
     database = getattr(mongo, MONGO_TEST_DATABASE)
     collection = getattr(database, MONGO_TEST_COLLECTION)
@@ -109,49 +125,134 @@ def test_backends_database_mongo_get_method_with_a_custom_query(mongo):
 def test_backends_database_mongo_to_documents_method():
     """Test the mongo backend to_documents method."""
 
-    statements = [{"id": "foo"}, {"id": "bar"}, {"id": "bar"}]
+    timestamp = {"timestamp": "2022-06-27T15:36:50"}
+    statements = [
+        {"id": "foo", **timestamp},
+        {"id": "bar", **timestamp},
+        {"id": "bar", **timestamp},
+    ]
     documents = MongoDatabase.to_documents(statements)
 
     assert next(documents) == {
-        "_id": ObjectId("2c26b46b68ffc68ff99b453c"),
-        "_source": {"id": "foo"},
+        "_id": ObjectId("62b9ce922c26b46b68ffc68f"),
+        "_source": {"id": "foo", **timestamp},
     }
     assert next(documents) == {
-        "_id": ObjectId("fcde2b2edba56bf408601fb7"),
-        "_source": {"id": "bar"},
+        "_id": ObjectId("62b9ce92fcde2b2edba56bf4"),
+        "_source": {"id": "bar", **timestamp},
     }
     # Identical statement ID produces the same ObjectId
     assert next(documents) == {
-        "_id": ObjectId("fcde2b2edba56bf408601fb7"),
-        "_source": {"id": "bar"},
+        "_id": ObjectId("62b9ce92fcde2b2edba56bf4"),
+        "_source": {"id": "bar", **timestamp},
     }
 
 
 def test_backends_database_mongo_to_documents_method_when_statement_has_no_id(caplog):
     """Test the mongo backend to_documents method when a statement has no id field."""
 
-    statements = [{"id": "foo"}, {}, {"id": "bar"}]
+    timestamp = {"timestamp": "2022-06-27T15:36:50"}
+    statements = [{"id": "foo", **timestamp}, timestamp, {"id": "bar", **timestamp}]
 
     documents = MongoDatabase.to_documents(statements, ignore_errors=False)
     assert next(documents) == {
-        "_id": ObjectId("2c26b46b68ffc68ff99b453c"),
-        "_source": {"id": "foo"},
+        "_id": ObjectId("62b9ce922c26b46b68ffc68f"),
+        "_source": {"id": "foo", **timestamp},
     }
-    with pytest.raises(BadFormatException, match="statement {} has no 'id' field"):
+    with pytest.raises(
+        BadFormatException, match=f"statement {timestamp} has no 'id' field"
+    ):
         next(documents)
 
     documents = MongoDatabase.to_documents(statements, ignore_errors=True)
     assert next(documents) == {
-        "_id": ObjectId("2c26b46b68ffc68ff99b453c"),
-        "_source": {"id": "foo"},
+        "_id": ObjectId("62b9ce922c26b46b68ffc68f"),
+        "_source": {"id": "foo", **timestamp},
     }
     assert next(documents) == {
-        "_id": ObjectId("fcde2b2edba56bf408601fb7"),
-        "_source": {"id": "bar"},
+        "_id": ObjectId("62b9ce92fcde2b2edba56bf4"),
+        "_source": {"id": "bar", **timestamp},
     }
     assert len(caplog.records) == 1
     assert caplog.records[0].levelname == "WARNING"
-    assert caplog.records[0].message == "statement {} has no 'id' field"
+    assert caplog.records[0].message == f"statement {timestamp} has no 'id' field"
+
+
+def test_backends_database_mongo_to_documents_method_when_statement_has_no_timestamp(
+    caplog,
+):
+    """Tests the mongo backend to_documents method when a statement has no timestamp."""
+
+    timestamp = {"timestamp": "2022-06-27T15:36:50"}
+    statements = [{"id": "foo", **timestamp}, {"id": "bar"}, {"id": "baz", **timestamp}]
+
+    documents = MongoDatabase.to_documents(statements, ignore_errors=False)
+    assert next(documents) == {
+        "_id": ObjectId("62b9ce922c26b46b68ffc68f"),
+        "_source": {"id": "foo", **timestamp},
+    }
+
+    with pytest.raises(
+        BadFormatException, match="statement {'id': 'bar'} has no 'timestamp' field"
+    ):
+        next(documents)
+
+    documents = MongoDatabase.to_documents(statements, ignore_errors=True)
+    assert next(documents) == {
+        "_id": ObjectId("62b9ce922c26b46b68ffc68f"),
+        "_source": {"id": "foo", **timestamp},
+    }
+    assert next(documents) == {
+        "_id": ObjectId("62b9ce92baa5a0964d3320fb"),
+        "_source": {"id": "baz", **timestamp},
+    }
+    assert len(caplog.records) == 1
+    assert caplog.records[0].levelname == "WARNING"
+    assert caplog.records[0].message == (
+        "statement {'id': 'bar'} has no 'timestamp' field"
+    )
+
+
+def test_backends_database_mongo_to_documents_method_with_invalid_timestamp(caplog):
+    """Tests the mongo backend to_documents method given a statement with an invalid
+    timestamp.
+    """
+
+    valid_timestamp = {"timestamp": "2022-06-27T15:36:50"}
+    invalid_timestamp = {"timestamp": "This is not a valid timestamp!"}
+    invalid_statement = {"id": "bar", **invalid_timestamp}
+    statements = [
+        {"id": "foo", **valid_timestamp},
+        invalid_statement,
+        {"id": "baz", **valid_timestamp},
+    ]
+
+    documents = MongoDatabase.to_documents(statements, ignore_errors=False)
+    assert next(documents) == {
+        "_id": ObjectId("62b9ce922c26b46b68ffc68f"),
+        "_source": {"id": "foo", **valid_timestamp},
+    }
+
+    with pytest.raises(
+        BadFormatException,
+        match=f"statement {invalid_statement} has an invalid 'timestamp' field",
+    ):
+        next(documents)
+
+    documents = MongoDatabase.to_documents(statements, ignore_errors=True)
+    assert next(documents) == {
+        "_id": ObjectId("62b9ce922c26b46b68ffc68f"),
+        "_source": {"id": "foo", **valid_timestamp},
+    }
+    assert next(documents) == {
+        "_id": ObjectId("62b9ce92baa5a0964d3320fb"),
+        "_source": {"id": "baz", **valid_timestamp},
+    }
+    assert len(caplog.records) == 1
+    assert caplog.records[0].levelname == "WARNING"
+    assert caplog.records[0].message == (
+        f"statement {invalid_statement} has an invalid 'timestamp' field"
+    )
 
 
 def test_backends_database_mongo_bulk_import_method(mongo):
@@ -163,17 +264,18 @@ def test_backends_database_mongo_bulk_import_method(mongo):
         database=MONGO_TEST_DATABASE,
         collection=MONGO_TEST_COLLECTION,
     )
-    statements = [{"id": "foo"}, {"id": "bar"}]
+    timestamp = {"timestamp": "2022-06-27T15:36:50"}
+    statements = [{"id": "foo", **timestamp}, {"id": "bar", **timestamp}]
     backend.bulk_import(MongoDatabase.to_documents(statements))
 
     results = backend.collection.find()
     assert next(results) == {
-        "_id": ObjectId("2c26b46b68ffc68ff99b453c"),
-        "_source": {"id": "foo"},
+        "_id": ObjectId("62b9ce922c26b46b68ffc68f"),
+        "_source": {"id": "foo", **timestamp},
     }
     assert next(results) == {
-        "_id": ObjectId("fcde2b2edba56bf408601fb7"),
-        "_source": {"id": "bar"},
+        "_id": ObjectId("62b9ce92fcde2b2edba56bf4"),
+        "_source": {"id": "bar", **timestamp},
     }
 
 
@@ -189,9 +291,14 @@ def test_backends_database_mongo_bulk_import_method_with_duplicated_key(mongo):
 
     # Identical statement ID produces the same ObjectId, leading to a
     # duplicated key write error while trying to bulk import this batch
-    statements = [{"id": "foo"}, {"id": "bar"}, {"id": "bar"}]
+    timestamp = {"timestamp": "2022-06-27T15:36:50"}
+    statements = [
+        {"id": "foo", **timestamp},
+        {"id": "bar", **timestamp},
+        {"id": "bar", **timestamp},
+    ]
     documents = list(MongoDatabase.to_documents(statements))
-    with pytest.raises(BulkWriteError, match="E11000 duplicate key error collection"):
+    with pytest.raises(BackendException, match="E11000 duplicate key error collection"):
         backend.bulk_import(documents)
 
     success = backend.bulk_import(documents, ignore_errors=True)
@@ -214,12 +321,13 @@ def test_backends_database_mongo_bulk_import_method_import_partial_chunks_on_err
 
     # Identical statement ID produces the same ObjectId, leading to a
     # duplicated key write error while trying to bulk import this batch
+    timestamp = {"timestamp": "2022-06-27T15:36:50"}
     statements = [
-        {"id": "foo"},
-        {"id": "bar"},
-        {"id": "baz"},
-        {"id": "bar"},
-        {"id": "lol"},
+        {"id": "foo", **timestamp},
+        {"id": "bar", **timestamp},
+        {"id": "baz", **timestamp},
+        {"id": "bar", **timestamp},
+        {"id": "lol", **timestamp},
     ]
     documents = list(MongoDatabase.to_documents(statements))
     assert backend.bulk_import(documents, ignore_errors=True) == 3
@@ -232,7 +340,8 @@ def test_backends_database_mongo_put_method(mongo):
     collection = getattr(database, MONGO_TEST_COLLECTION)
     assert collection.estimated_document_count() == 0
 
-    statements = [{"id": "foo"}, {"id": "bar"}]
+    timestamp = {"timestamp": "2022-06-27T15:36:50"}
+    statements = [{"id": "foo", **timestamp}, {"id": "bar", **timestamp}]
     backend = MongoDatabase(
         connection_uri=MONGO_TEST_URI,
         database=MONGO_TEST_DATABASE,
@@ -245,12 +354,12 @@ def test_backends_database_mongo_put_method(mongo):
 
     results = collection.find()
     assert next(results) == {
-        "_id": ObjectId("2c26b46b68ffc68ff99b453c"),
-        "_source": {"id": "foo"},
+        "_id": ObjectId("62b9ce922c26b46b68ffc68f"),
+        "_source": {"id": "foo", **timestamp},
     }
     assert next(results) == {
-        "_id": ObjectId("fcde2b2edba56bf408601fb7"),
-        "_source": {"id": "bar"},
+        "_id": ObjectId("62b9ce92fcde2b2edba56bf4"),
+        "_source": {"id": "bar", **timestamp},
     }
 
 
@@ -261,7 +370,8 @@ def test_backends_database_mongo_put_method_with_custom_chunk_size(mongo):
     collection = getattr(database, MONGO_TEST_COLLECTION)
     assert collection.estimated_document_count() == 0
 
-    statements = [{"id": "foo"}, {"id": "bar"}]
+    timestamp = {"timestamp": "2022-06-27T15:36:50"}
+    statements = [{"id": "foo", **timestamp}, {"id": "bar", **timestamp}]
     backend = MongoDatabase(
         connection_uri=MONGO_TEST_URI,
         database=MONGO_TEST_DATABASE,
@@ -274,10 +384,10 @@ def test_backends_database_mongo_put_method_with_custom_chunk_size(mongo):
 
     results = collection.find()
     assert next(results) == {
-        "_id": ObjectId("2c26b46b68ffc68ff99b453c"),
-        "_source": {"id": "foo"},
+        "_id": ObjectId("62b9ce922c26b46b68ffc68f"),
+        "_source": {"id": "foo", **timestamp},
     }
     assert next(results) == {
-        "_id": ObjectId("fcde2b2edba56bf408601fb7"),
-        "_source": {"id": "bar"},
+        "_id": ObjectId("62b9ce92fcde2b2edba56bf4"),
+        "_source": {"id": "bar", **timestamp},
     }